+++ src/main/java/IDMaker/idmaker/IDMaker.java
@@ -1,4 +1,4 @@
-package IDMaker.project;
+package IDMaker.idmaker;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
