+++ src/main/java/IDMaker/idmaker/IDMakerProject.java
@@ -1,4 +1,4 @@
-package IDMaker.project;
+package IDMaker.idmaker;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
