+++ src/main/java/IDMaker/idmaker/RandomValueUtil.java
@@ -1,4 +1,4 @@
-package IDMaker.project;
+package IDMaker.idmaker;
 
 import java.util.Random;
