+++ src/main/java/IDMaker/idmaker/fixture/controller/TestClassController.java
@@ -1,12 +1,12 @@
-package IDMaker.project.fixture.controller;
+package IDMaker.idmaker.fixture.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import IDMaker.project.fixture.entity.TestClass;
-import IDMaker.project.fixture.service.TestClassService;
+import IDMaker.idmaker.fixture.entity.TestClass;
+import IDMaker.idmaker.fixture.service.TestClassService;
 
 @RestController
 public class TestClassController {
