+++ src/main/java/IDMaker/idmaker/fixture/repository/TestClassRepository.java
@@ -1,8 +1,8 @@
-package IDMaker.project.fixture.repository;
+package IDMaker.idmaker.fixture.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import IDMaker.project.fixture.entity.TestClass;
+import IDMaker.idmaker.fixture.entity.TestClass;
 
 public interface TestClassRepository extends JpaRepository<TestClass, String> {
 }
