+++ src/test/java/IDMaker/project/test/IDMakerTest.java
@@ -12,10 +12,10 @@ import java.util.stream.IntStream;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 
-import IDMaker.project.GenerationType;
-import IDMaker.project.fixture.entity.TestClass;
-import IDMaker.project.fixture.service.TestClassService;
-import IDMaker.project.IDMakerUtils;
+import IDMaker.idmaker.GenerationType;
+import IDMaker.idmaker.fixture.entity.TestClass;
+import IDMaker.idmaker.fixture.service.TestClassService;
+import IDMaker.idmaker.IDMakerUtils;
 
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
